+++ CHANGELOG.md
@@ -5,6 +5,16 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [1.18.2] - 2023-04-19
+
+### Changed
+
+- Moved delete user to logic class
+
+### Added
+
+- Command for deleting user manually
+
 ## [1.18.1] - 2023-04-18
 
 ### Added